refactor structure files & delete templates codes
